+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/SnapshotImpl.java
@@ -42,8 +42,8 @@ import io.delta.kernel.internal.replay.LogReplay;
 import io.delta.kernel.internal.snapshot.LogSegment;
 import io.delta.kernel.internal.util.VectorUtils;
 import io.delta.kernel.metrics.SnapshotReport;
-import io.delta.kernel.transaction.UpdateTableTransactionBuilder;
 import io.delta.kernel.transaction.ReplaceTableTransactionBuilder;
+import io.delta.kernel.transaction.UpdateTableTransactionBuilder;
 import io.delta.kernel.types.StructType;
 import java.util.List;
 import java.util.Map;
